+++ src/main/java/FIS/iLUVit/service/CenterService.java
@@ -48,7 +48,7 @@ public class CenterService {
                 centerRepository.findByFilterForMapList(longitude, latitude, userId, kindOf, centerIds, pageable);
     }
 
-    public List<CenterMapPreview> findByFilterForMap(double longitude, double latitude, Integer distance, String searchContent){
+    public List<CenterMapPreview> findByFilterForMap(double longitude, double latitude, Double distance, String searchContent){
         return centerRepository.findByFilterForMap(longitude, latitude, distance, searchContent);
     }
